HostMaster: move AgentProcessProtocol into ocs lib
And also rename the Docker one -- this anticipates smoothing their
APIs together to simplify HostMaster code.
